deploy_aws/requirements.txt
google-generativeai>=0.3.0
pandas>=1.5.0
openpyxl>=3.0.0
schedule>=1.2.0
requests>=2.28.0
flask>=2.3.0
